feat(formatter): output parent id
fix #28
