+++ packages/@secretlint/formatter/test/snapshots/input.no-secret.txt
@@ -1,0 +1,1 @@
+THIS IS SECRET SOURCE
