+++ packages/@secretlint/formatter/test/snapshots/input.ts
@@ -1,0 +1,72 @@
+import { SecretLintCoreResult } from "@secretlint/types";
+import path from "path";
+
+export const results: SecretLintCoreResult[] = [
+    {
+        filePath: path.join(__dirname, "input.txt"),
+        messages: [
+            {
+                data: {
+                    ID: "SECRET"
+                },
+                message: "warning found secret: SECRET",
+                range: [8, 14],
+                loc: {
+                    start: {
+                        line: 1,
+                        column: 8
+                    },
+                    end: {
+                        line: 1,
+                        column: 14
+                    }
+                },
+                ruleId: "example",
+                severity: "warning"
+            },
+            {
+                data: {
+                    ID: "SECRET"
+                },
+                message: "error found secret: SECRET",
+                range: [8, 14],
+                loc: {
+                    start: {
+                        line: 1,
+                        column: 8
+                    },
+                    end: {
+                        line: 1,
+                        column: 14
+                    }
+                },
+                ruleId: "example",
+                severity: "error"
+            },
+            {
+                data: {
+                    ID: "SECRET"
+                },
+                message: "error found secret: SECRET",
+                range: [8, 14],
+                loc: {
+                    start: {
+                        line: 1,
+                        column: 8
+                    },
+                    end: {
+                        line: 1,
+                        column: 14
+                    }
+                },
+                ruleId: "example",
+                ruleParentId: "parent",
+                severity: "error"
+            }
+        ]
+    },
+    {
+        filePath: path.join(__dirname, "input.no-secret.txt"),
+        messages: []
+    }
+];
