+++ packages/@secretlint/secretlint-rule-preset-recommend/test/snapshots/aws/output.json
@@ -8,6 +8,7 @@
                 32
             ],
             "ruleId": "@secretlint/secretlint-rule-aws",
+            "ruleParentId": "@secretlint/secretlint-rule-preset-recommend",
             "loc": {
                 "start": {
                     "line": 1,
